feat: support long-lived connection between the background and the UI.
- Update `pushMessage()` and `postMessage`
- Split `ConfirmRequestModal` into different kinds of smaller modals
- Creates a new `approves` duck script
- Update `README` title

Refs: #75
