+++ src/background/controllers/requestManager.ts
@@ -15,7 +15,7 @@ export default class RequestManager extends EventEmitter2 {
     this.pendingRequests = [];
   }
 
-  getRequests = (): PendingRequest[] => this.pendingRequests;
+  getRequests = () => pushMessage(setPendingRequest(this.pendingRequests));
 
   finalizeRequest = async (action: RequestResolutionAction<any>): Promise<boolean> => {
     const { id } = action;
@@ -23,7 +23,7 @@ export default class RequestManager extends EventEmitter2 {
     // TODO add some mutex lock just in case something strange occurs
     this.pendingRequests = this.pendingRequests.filter(pendingRequest => pendingRequest.id !== id);
     this.emit(`${id}:finalized`, action);
-    await pushMessage(setPendingRequest(this.pendingRequests));
+    pushMessage(setPendingRequest(this.pendingRequests));
     return true;
   };
 
@@ -31,7 +31,7 @@ export default class RequestManager extends EventEmitter2 {
     // eslint-disable-next-line no-plusplus
     const id = `${nonce++}`;
     this.pendingRequests.push({ id, type, payload });
-    await pushMessage(setPendingRequest(this.pendingRequests));
+    pushMessage(setPendingRequest(this.pendingRequests));
 
     return id;
   };
