+++ src/ui/ducks/requests.ts
@@ -1,21 +1,13 @@
 import { useSelector } from "react-redux";
 import { AppRootState } from "@src/ui/store/configureAppStore";
 import deepEqual from "fast-deep-equal";
-import { PendingRequest } from "@src/types";
-import { Dispatch } from "redux";
+import { Action, PendingRequest } from "@src/types";
 import RPCAction from "@src/util/constants";
 import postMessage from "@src/util/postMessage";
 
 enum ActionType {
   SET_PENDING_REQUESTS = "request/setPendingRequests",
 }
-
-type Action<payload> = {
-  type: ActionType;
-  payload?: payload;
-  meta?: any;
-  error?: boolean;
-};
 
 type State = {
   pendingRequests: PendingRequest[];
@@ -25,17 +17,16 @@ const initialState: State = {
   pendingRequests: [],
 };
 
-export const setPendingRequest = (pendingRequests: PendingRequest[]): Action<PendingRequest[]> => ({
+export const setPendingRequest = (pendingRequests: PendingRequest[]): Action<ActionType, PendingRequest[]> => ({
   type: ActionType.SET_PENDING_REQUESTS,
   payload: pendingRequests,
 });
 
-export const fetchRequestPendingStatus = () => async (dispatch: Dispatch) => {
-  const pendingRequests = await postMessage({ method: RPCAction.GET_PENDING_REQUESTS });
-  dispatch(setPendingRequest(pendingRequests));
+export const fetchRequestPendingStatus = () => async () => {
+  postMessage({ method: RPCAction.GET_PENDING_REQUESTS });
 };
 
-export default function requests(state = initialState, action: Action<any>): State {
+export default function requests(state = initialState, action: Action<ActionType, any>): State {
   switch (action.type) {
     case ActionType.SET_PENDING_REQUESTS:
       return {
