+++ src/util/postMessage.ts
@@ -1,13 +1,32 @@
-import { browser } from "webextension-polyfill-ts";
-
 import type { MessageAction } from "@src/types";
 
-export default async function postMessage(message: MessageAction) {
-  const [err, res] = await browser.runtime.sendMessage(message);
+import log from "loglevel";
 
-  if (err) {
-    throw new Error(err);
+import { Runtime } from "webextension-polyfill";
+
+let messageSender: MessageSender;
+
+class MessageSender {
+  public remotePort: Runtime.Port;
+
+  constructor(remotePort: Runtime.Port) {
+    this.remotePort = remotePort;
   }
 
-  return res;
+  public send(message: MessageAction) {
+    try {
+      this.remotePort.postMessage(message);
+    } catch (error) {
+      log.warn("Push message error: ", error);
+    }
+  }
+}
+
+export function messageSenderFactory(remotePort: Runtime.Port) {
+  messageSender = new MessageSender(remotePort);
+  Object.freeze(messageSender);
+}
+
+export default function postMessage(message: MessageAction) {
+  messageSender.send(message);
 }
