+++ src/util/pushMessage.ts
@@ -1,11 +1,34 @@
-import { browser } from "webextension-polyfill-ts";
 import log from "loglevel";
 
 import { ReduxAction } from "@src/types";
+import { Runtime } from "webextension-polyfill";
+
+let messageSender: MessageSender;
+
+class MessageSender {
+  public remotePort: Runtime.Port;
+
+  constructor(remotePort: Runtime.Port) {
+    this.remotePort = remotePort;
+  }
+
+  public send(message: ReduxAction) {
+    try {
+      this.remotePort.postMessage(message);
+    } catch (error) {
+      log.warn("Push message error: ", error);
+    }
+  }
+}
+
+export function messageSenderFactory(remotePort: Runtime.Port) {
+  messageSender = new MessageSender(remotePort);
+  Object.freeze(messageSender);
+}
 
 export default async function pushMessage(message: ReduxAction) {
   try {
-    await browser.runtime.sendMessage(message);
+    messageSender.send(message);
   } catch (error) {
     log.warn("Push message error: ", error);
   }
